+++ src/app.module.ts
@@ -1,9 +1,10 @@
+import { CryptographyModule } from '@app/cryptography';
 import { EnvModule } from '@env/env';
 import { envSchema } from '@env/env/env';
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AuthModule } from './modules/auth/auth.module';
-import { CryptographyModule } from './modules/cryptography/cryptography.module';
+import { ShortenerModule } from './modules/shortener/shortener.module';
 import { UsersModule } from './modules/users/users.module';
 
 @Module({
@@ -16,6 +17,7 @@ import { UsersModule } from './modules/users/users.module';
     CryptographyModule,
     UsersModule,
     AuthModule,
+    ShortenerModule,
   ],
   controllers: [],
   providers: [],